DXE-1251 clean up upgrade command
