+++ pkg/commands/command_upgrade.go
@@ -1,3 +1,6 @@
+//go:build !noautoupgrade
+// +build !noautoupgrade
+
 // Copyright 2018. Akamai Technologies, Inc
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
